### 2021-08-20 #3637 @whedon
[PRE REVIEW]: infer: An R package for tidyverse-friendly statistical inference
**Submitting author:** @simonpcouch (<a href="http://orcid.org/0000-0001-5676-5107">Simon Couch</a>)
**Repository:** <a href="https://github.com/tidymodels/infer" target ="_blank">https://github.com/tidymodels/infer</a>
**Version:** v1.0.0
