### 2021-08-20 #3637 @simonpcouch
Thanks, yall!

Some folks from the spreadsheet of reviewers that have listed relevant expertise: Kellie Ottoboni (at kellieotto),
Gjalt-Jorn Peters (at Matherion), Mine Dogucu (at mdogucu), and Pedro Rafael (at prdm0).

I'll get to work on addressing the licensing issue!
