### 2021-08-20 #3637 @simonpcouch
Okay, looks like the licensing is good to go—the CC0 license was previously only listed in the `DESCRIPTION` file, and now also lives in its own separate `LICENSE` file as well.
