### 2021-08-20 #3637 @danielskatz
Sorry, this is a problem. JOSS requires use of an [OSI-approved license](https://opensource.org/licenses) - this specifically [does not include CC0](https://opensource.org/faq#cc-zero).
