### 2021-08-20 #3637 @simonpcouch
Ah, thank you for the clarification. Beginning discussions with co-authors on licensing at the moment.
