### 2021-08-21 fecb951
note re-licensing in `NEWS.md`
