### 2021-08-21 #3637 @simonpcouch
Thanks again for the note—with agreement from all package contributors, we've re-licensed from CC0 to MIT. Documentation at tidymodels/infer#413!
