### 2021-08-22 #3637 @csoneson
👋🏻 @kellieotto, @mdogucu, @prdm0 - would (two of) you be interested in reviewing this submission for JOSS?

> infer: An R package for tidyverse-friendly statistical inference
