### 2021-09-06 tidymodels/infer#415 @kellieotto
[JOSS Review] Don't understand when generate(..., type="draw") is applicable
This comment is part of your [JOSS Review](https://github.com/openjournals/joss-reviews/issues/3661).

I'm unclear on how to use type="draw" in the `generate` function. At a minimum, it would be helpful to clarify the draw method and show an example how to use this in the documentation for `generate`.

From reading the docs, it seems like I should be able to use the draw method to sample data according to a point estimate null hypothesis. If I express that the null point estimate is mu=40, I expect that the draw method will randomly sample some normally distributed data with mean 40 and some variance. However this example gives an error:

```
> gss %>%
+     specify(response = hours) %>%
+     hypothesize(null = "point", mu = 40) %>%
+     generate(reps = 1000, type = "draw")
Error in sample.int(length(x), size, replace, prob) :
  NA in probability vector
In addition: Warning message:
You have given `type = "draw"`, but `type` is expected to be `"bootstrap"`. This workflow is untested and the results may not mean what you think they mean.
```

I see there are some examples where the draw method applies to proportions in `vignette("observed_stat_examples")`. If it only works with proportions, can you please clarify that in the `generate` docs and add an example there?
