### 2021-09-06 tidymodels/infer#416 @kellieotto
[JOSS Review] Random seeds for reproducibility
This comment is part of your [JOSS Review](https://github.com/openjournals/joss-reviews/issues/3661).

I see that there are random seeds set in the unit tests and it is mentioned in the GitHub issue prompt, but the importance of setting the random seed isn't mentioned in the documentation for the package. I think this is a crucial piece of documentation/information that is missing, especially since this a goal of this package is to be used for statistics pedagogy. As is, the user will get different results each time they run the code if they do not first set the seed.

Apart from including set.seed in all of the examples, you may want to consider how you could build it into the package, as a function argument or something that's stated in a series of functions.
