### 2021-09-06 #3661 @kellieotto
@csoneson @simonpcouch: I've gone through the review checklist and just left two issues pertaining to functionality documentation, https://github.com/tidymodels/infer/issues/415 and https://github.com/tidymodels/infer/issues/416. I'll take another look when those are resolved.
