### 2021-09-07 b719143
add `man-roxygen` section on `set.seed`
