### 2021-09-07 de02516
update wording with Mine's suggestion
