+++ man-roxygen/seeds.Rmd
@@ -16,7 +16,7 @@ gss %>%
   calculate("diff in means", order = c("degree", "no degree"))
 ```
 
-Setting the seed again and rerunning the same code will result in the same answer.
+Setting the seed to the same value again and rerunning the same code will produce the same result.
 
 ```{r}
 # set the seed
