### 2021-09-07 d93daa0
note `set.seed` in intro to infer vignette
