### 2021-09-07 #3661 @simonpcouch
Okay, addressed in tidymodels/infer#418 and tidymodels/infer#417, respectively. Thanks again!
