### 2021-09-08 tidymodels/infer#420 @prdm0
[JOSS Review] Review of the paper submitted to JOSS @prdm0
Essa é minha revisão para o paper submetido na JOSS, [#3661](https://github.com/openjournals/joss-reviews/issues/3661).

Dear @simonpcouch, I'm glad to be able to contribute to the **infer** package review. I'm a theoretical statistician but I get into programming a little bit in R. I love the metaprogramming capability of R and the great flexibility this language brings. I also learned to like the tidyverse programming style.

It would be very interesting for the **infer** package to allow returning measures of information in univariate distributions, which is easy to implement. A design for this would be:

1. You would pass a univariate cumulative distribution function and a data set, for example, `univariate_information_theory(G, data, ...)`. `G` can be generic, as every day a new distribution appears;
2.  Using the vectorization characteristic of the R language, it is possible to calculate the maximum likelihood estimators of the parameters that index `G`;
3. With these estimators, it is possible to calculate some traditional fit adequacy measures such as AIC, BIC, CAIC, HQIC (Hannan-Quinn information criterion), in addition to two other measures that are easy to be implemented and serve for discrete and continuous distributions , which are the Cramér-von Misses and Anderson-Darling measures modified by Gemain Chen and N. Balakrishnan (1995).

In the paper *The General Purpose Approximate Goodness-of-Fit Test* they present a very simple algorithm to implement these statistics which are widely referenced in papers in the area of ​​probability distributions and would fit very well with the proposal of the **infer** package . In the algorithm, you could numerically derive `G` to obtain the log-likelihood function and would have the advantage of not having to numerically integrate the density function `g`, since the Cramèr-von Misses statistics (W) and Anderson-Darling (A) modified by Gemain Chen and N. Balakrishnan use the `G` distribution. You would have to benchmark and feel the computational cost of numerically integrating or numerically deriving and deciding on the most efficient.

Note that the theoretical definition of statistics (A^*) and (W^*) uses a Lebesgue Stieltjes integral, which can be used in the discrete and continuous case, giving great visibility and usability to this function for the **infer** package. However, it is quite easy to implement these statistics as there is a very simple algorithm described in the paper.

I have a package (https://github.com/prdm0/AdequacyModel) that implements these statistics, from the time I was very amateur in R. Today I would certainly implement it in another way. I'm sure you can do the same with the tidyverse outfit.
