### 2021-09-08 #3661 @prdm0
Dear @csoneson and @simonpcouch, at issue [#420](https://github.com/tidymodels/infer/issues/420) I leave my review for the paper. Beforehand I accept the paper, but I would be happy to see at least a prototype of my suggestion implemented. Best regards.
