### 2021-09-09 tidymodels/infer#420 @simonpcouch
Thanks for your review, @prdm0! The functionality you describe seems really useful, and very much worth implementing in an R package. However, I’m not sure that {infer} is the right place for this functionality to reside.

A few thoughts here:
* One of the main advantages to {infer}’s approach is it’s abstraction. The majority of the package’s functionality lives within four or five verbs, which can be arranged and rearranged to explore a large array of hypotheses and statistical questions. The fact that supporting this functionality would require its own verb (and one whose relationship with the other core verbs is not immediately clear to me) is a sign to me that this functionality may be better fitted for a follow-up package to {infer}.
* We have discussed, at times, extensions of the package to calculate arbitrary statistics. See #196 and #175, for example. Generally, we’ve concluded that a limitation of {infer}’s design framework is that the suite of test statistics it can support is limited to what we can implement and maintain explicit support for ourselves; extensions to arbitrary statistics or other information measures are beyond the scope of the package.
* The package is intended for use in teaching introductory statistics courses and research contexts with simple experimental designs. The functionality you suggest, while a valuable next step in extending a grammar like {infer}’s, is outside of the scope of the package.

Thanks again for your thoughtful suggestion.
