### 2021-09-09 tidymodels/infer#420 @prdm0
Dear @simonpcouch , thanks for your reply.

I understand your arguments and even before them I signaled the acceptance of your work. I find the package useful for people who teach statistical inference or who want to perform simple / basic statistical inference procedures.

In many basic statistics courses, after a brief and not in depth introduction to statistical inference, many students wonder how to decide between an f_1 and f_2 distribution, for example. This is one of the most common questions in a Basic Inference course.

The Cramèr-von Misses (W) and Anderson-Darling (A) statistics modified by Gemain Chen and N. Balakrishnan are methods almost as simple as an X^2 test, with no asymptotic dependence for a theoretical distribution. So a teacher would be comfortable telling their student to use **infer** package to do a simple X^2 test when some strong conditions are met, or even use **infer** to understand that there are other test statistics that work well on small samples, when strong conditions are not met (which is more common). The student would understand without great theoretical details that besides the X^2 method there are other more sophisticated methodologies, but not extremely complex, to decide on the choice of some compared distribution.

I think the package would in no way lose its principle of parsimony. On the contrary, it would help the teacher to show the student that in situations where the X^2 test does not work, there are others that are very simple to implement and that are very informative, even in small samples.

I think progressing on some basic statistical methodologies would be a natural path, as the basics are still well beyond the implemented.

Finally, I ratify the quality of your package and I remain congruent with the acceptance of the paper.
