### 2021-09-11 #3661 @csoneson
👋🏻 @simonpcouch - I went through the submission and it looks good to me! Only a couple of things related to the references:

* There's a missing space between Modern and Dive in Ismay & Kim (2019)
* Could you provide more info/a link for Downey (2016)?
* I'd recommend adding references for the packages mentioned in the "Comparison to other packages" section

Could you fix these and ask whedon to generate a new proof (with @whedon generate pdf)?

The next steps would then be to:
- [ ] Make a tagged release of the software, and list the version tag of the archived version here.
- [ ] Check that the archival deposit (e.g., in Zenodo) has the correct metadata. In particular, the title and author list should be identical to those of the paper. You can also add the authors' ORCID.
