### 2021-09-12 a147f14
refine JOSS paper citations
