+++ figs/paper/paper.bib
@@ -118,5 +118,49 @@
   title={There is still only one test},
   author={Downey, Allen B},
   year={2016},
-  publisher={"Probably Overthinking It"}
+  publisher={"Probably Overthinking It"},
+  url = {http://allendowney.blogspot.com/2016/06/there-is-still-only-one-test.html}
+}
+
+@Article{r-mosaic,
+  author = {Randall Pruim and Daniel T Kaplan and Nicholas J Horton},
+  title = {The mosaic Package: Helping Students to 'Think with Data' Using R},
+  journal = {The R Journal},
+  volume = {9},
+  number = {1},
+  pages = {77--102},
+  year = {2017},
+  url = {https://journal.r-project.org/archive/2017/RJ-2017-024/index.html},
+}
+
+@Article{r-statsExpressions,
+  doi = {10.21105/joss.03236},
+  url = {https://doi.org/10.21105/joss.03236},
+  year = {2021},
+  publisher = {{The Open Journal}},
+  volume = {6},
+  number = {61},
+  pages = {3236},
+  author = {Indrajeet Patil},
+  title = {{statsExpressions: {R} Package for Tidy Dataframes and Expressions with Statistical Details}},
+  journal = {{Journal of Open Source Software}},
+}
+
+@Article{r-parameters,
+  title = {Extracting, Computing and Exploring the Parameters of Statistical Models using {R}.},
+  volume = {5},
+  doi = {10.21105/joss.02445},
+  number = {53},
+  journal = {Journal of Open Source Software},
+  author = {Daniel Lüdecke and Mattan S. Ben-Shachar and Indrajeet Patil and Dominique Makowski},
+  year = {2020},
+  pages = {2445},
+}
+  
+@Manual{r-broom,
+  title = {broom: Convert Statistical Objects into Tidy Tibbles},
+  author = {David Robinson and Alex Hayes and Simon Couch},
+  year = {2021},
+  note = {R package version 0.7.9},
+  url = {https://CRAN.R-project.org/package=broom},
 }
