### 2021-09-12 #3661 @simonpcouch
Thank you, @csoneson! I've made each of the changes you've suggested, with one exception: the authors of the book do indeed spell ModernDive [without a space](https://moderndive.com/).

I just reached out to the admins of the tidymodels GitHub org to enable Zenodo's access to `tidymodels/infer`. I'll drop a note here once we've made it through each of the steps you've listed to generate an archive + DOI!
