### 2021-09-13 be84547
bump version for JOSS release
