+++ NEWS.md
@@ -1,4 +1,6 @@
-# infer v1.0.0.9000 (development version)
+# infer v1.0.1 (GitHub Only)
+
+This release reflects the infer version accepted to the Journal of Open Source Software.
 
 * Re-licensed the package from CC0 to MIT. See the `LICENSE` and `LICENSE.md` files.
 * Contributed a paper to the Journal of Open Source Software, a draft of which is available in `/figs/paper`.
