+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: infer
 Title: Tidy Statistical Inference
-Version: 1.0.0.9000
+Version: 1.0.1
 Authors@R:
     c(person(given = "Andrew",
              family = "Bray",
