### 2021-09-13 #3661 @simonpcouch
Okay, the DOI should be good to go! The DOI of the archived version, with correct metadata, is here: 10.5281/zenodo.5504495
