### 2021-09-13 #3661 @simonpcouch
Ah, and the version tag is `v1.0.1`.
