+++ figs/paper/paper.bib
@@ -124,7 +124,7 @@
 
 @Article{r-mosaic,
   author = {Randall Pruim and Daniel T Kaplan and Nicholas J Horton},
-  title = {The mosaic Package: Helping Students to 'Think with Data' Using R},
+  title = {The mosaic Package: Helping Students to 'Think with Data' Using {R}},
   journal = {The R Journal},
   volume = {9},
   number = {1},
