+++ figs/paper/paper.bib
@@ -86,7 +86,7 @@
 }
 
 @article{kuhn2020tidymodels,
-  title={Tidymodels: a collection of packages for modeling and machine learning using tidyverse principles},
+  title={{tidymodels}: a collection of packages for modeling and machine learning using tidyverse principles},
   author={Kuhn, M and Wickham, H},
   journal={Boston, MA, USA.},
   year={2020},
@@ -124,7 +124,7 @@
 
 @Article{r-mosaic,
   author = {Randall Pruim and Daniel T Kaplan and Nicholas J Horton},
-  title = {The mosaic Package: Helping Students to 'Think with Data' Using {R}},
+  title = {The mosaic Package: Helping Students to `Think with Data' Using {R}},
   journal = {The R Journal},
   volume = {9},
   number = {1},
@@ -158,7 +158,7 @@
 }
   
 @Manual{r-broom,
-  title = {broom: Convert Statistical Objects into Tidy Tibbles},
+  title = {{broom}: Convert Statistical Objects into Tidy Tibbles},
   author = {David Robinson and Alex Hayes and Simon Couch},
   year = {2021},
   note = {R package version 0.7.9},
