+++ figs/paper/paper.bib
@@ -130,7 +130,7 @@
   number = {1},
   pages = {77--102},
   year = {2017},
-  url = {https://journal.r-project.org/archive/2017/RJ-2017-024/index.html},
+  doi = {10.32614/RJ-2017-024},
 }
 
 @Article{r-statsExpressions,
