### 2021-09-15 #3661 @simonpcouch
Ah, that new flag for the missing `mosaic` DOI is correct. Just fixed!
