### 2021-09-16 #3661 @Kevin-Mattheus-Moerman
@simonpcouch congratulations on getting this work published in JOSS

@csoneson, thanks for editing!

@prdm0, @kellieotto thanks a lot for your help reviewing this work!
